pyproject.toml
[tool.poetry]
name = "database-paper-graph"
version = "0.1.0"
description = ""
authors = ["TatsukiNishimura <hbvcg00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
pytest = "^8.3.3"
django-cors-headers = "^4.6.0"


[tool.poetry.group.isort.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
